requirements.txt
flask
flask-cors
pyserial
adafruit-circuitpython-bme680
